feat(shift-typeahead): add on select/deselect hooks
I would've loved to just pass the onOptionSelect down to shift-multiselect but angular won't let that happen.. I'm unable to pass a callback reference  ('&' binding) down to another callback
reference in a child directive. The workaround was to use these intermediate methods defined on the scope: onMultiOptionSelect and onMultiOptionDeselect which will invoke the callback reference provided to shift-typeahead via onOptionSelect/onOptionDeselect.