fix(shift-typeahead): proceed only when menu is visible
shift-typeahead will take control of the ESC key when close_menu_on_esc is `true`. this is to avoid conflicting behavior with other listeners of the ESC key. For example, a modal that listens on the document for an ESC keypress to close the modal would then close the modal when the select menu is open instead of closing the menu.